// apps/qa/src/app/qa/feature/qa-sticker-flow/data-access/qa-sticker-api.ts
export namespace QaStickerScanSheetApi {
  export type Response = {
    orderId: number;
    // partnerId: number;
    customPartnerId: string;
    xid: string;
    slaDateOnUtc: Date;
    xqc: boolean;
    isPriority: boolean;
    shippingAlert: string | null;
    isVipOrder: boolean;
    sheetBarcode: string;
    stickerSheetId: number;
    qaLeadRejectReviewBarcodes: QaLeadRejectReviewBarcode[];
    coos: {
      id: number;
      name: string;
      code: string;
      cooImageUrl: string | null;
      transformName: string;
    }[];

    //region flow related
    coo: Response['coos'][number] | null;
    isAllowChangeCoo: boolean;
    status: ItemTransitionStatus;
    statusDescription: string;
    //endregion

    steps: (QaStickerStepStickerGroup | QaStickerStepCoo | QaStickerStepPackage | QaStickerStepTrackingNumber | QaStickerStepFinal)[];
    isJit: boolean;
  };

  export type BaseStep = {
    isIgnoreScan: boolean;
    isViewOnly: boolean;
    isScanned: boolean;
  };

  export type QaStickerStepStickerGroup = BaseStep & {
    groupType: QaStepGroup.StickerGroup;
    scanningStickerGroup: {
      barcode: string;
      // status: ItemTransitionStatus;
      // sku: string;
      styleDesc: string;
      colorName: string;
      sizeCode: string;
      sizeClassName: string;
      rejectedCount: number;
      isRejectWithoutConfirming: boolean; // qa lead
      // isQaSuccessAll: boolean;
      isAllowChangeCoo: boolean; // move this outside for easier usage
      customNumber: string | null;
      customName: string | null;
      attribute: {
        orderDetailAttributeId: number;
        // isNoDesign: boolean;
        isReadyForQa: boolean;
        locationCode: string;
        locationName: string;
        description: string | null;
        locationStatus: ItemTransitionStatus;
        fileUrl: string;
        scanAction: QaPrintScanAction | null;
        reviewBarcodes: {
          location: QaPrintBarcodeLocation;
          type: QaPrintScanAction;
          barcode: string;
        }[];
      };
    };
  };
  export type QaStickerStepCoo = BaseStep & {
    groupType: QaStepGroup.Coo;
  };
  export type QaStickerStepPackage = BaseStep & {
    groupType: QaStepGroup.Package;
    scanningPackages: {
      scannedPackage: QaStickerStepPackage['scanningPackages']['packages'][number] | null;
      packages: {
        id: number;
        name: string;
        barcode: string;
        imageUrl: string;
      }[];
    };
  };
  export type QaStickerStepTrackingNumber = BaseStep & {
    groupType: QaStepGroup.TrackingNumber;
    scanningTrackingNumber: {
      trackingNumber: string;
    };
  };
  export type QaStickerStepFinal = BaseStep & {
    groupType: QaStepGroup.Final;
    message: string;
    messageColor: QaColor;
    finalStepMessageParams: {
      [k: string]: string | number;
    };
    messages?: {
      message: string;
      messageColor: QaColor;
    }[];
  };
}